Maintain #include interface for ease of use
